backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database drivers
psycopg2-binary==2.9.9
neo4j==5.16.0
redis==5.0.1

# Cloud providers
google-cloud-storage==2.14.0
google-cloud-kms==2.16.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# LLM and AI
openai==1.10.0
langchain==0.1.4
langchain-openai==0.0.5
langgraph==0.0.20

# HTTP client
httpx==0.26.0
aiohttp==3.9.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Development
black==23.12.1
flake8==7.0.0
mypy==1.8.0
